// src/store/user/user.types.ts
export enum USER_ACTION_TYPES {
  CHECK_USER_SESSION = "user/CHECK_USER_SESSION",
  SIGN_IN_WITH_EMAIL_START = "user/SIGN_IN_WITH_EMAIL_START",
  SIGN_IN_WITH_GOOGLE_START = "user/SIGN_IN_WITH_GOOGLE_START",
  SIGN_UP_WITH_EMAIL_START = "user/SIGN_UP_WITH_EMAIL_START",
  SIGN_IN_SUCCESS = "user/SIGN_IN_SUCCESS",
  SIGN_IN_FAILED = "user/SIGN_IN_FAILED",
  SIGN_OUT_START = "user/SIGN_OUT_START",
  SIGN_OUT_SUCCESS = "user/SIGN_OUT_SUCCESS",
  SIGN_OUT_FAILED = "user/SIGN_OUT_FAILED"
}